Refactor App component to use UserSidebar and AdminSidebar; update CreateIncident import path; fix lint error in layout; adjust LandingPage to import EmergencyContactCard
