+++ client/src/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { Provider } from 'react-redux'; // Importing Provider for Redux
 import { store } from "./store/store"; // Importing the Redux store
 import HomePage from "./pages/HomePage";
-import CreateIncident from "./pages/CreateIncident";
+import CreateIncident from "./components/CreateIncident";
 import Navbar from './components/Navbar';
 import Header from './components/Header';
 import LandingPage from './pages/LandingPage';
@@ -11,7 +11,8 @@ import Footer from "./components/Footer";
 import AdminDashboard from './pages/AdminDashboard'; // Ensure AdminDashboard is imported
 import ContactPage from './pages/ContactPage'; // Ensure ContactPage is imported
 import ManageIncidents from './pages/ManageIncidents';
-import Sidebar from './components/Sidebar'; // Ensure Sidebar is imported
+import UserSidebar from "./components/UserSidebar";
+import AdminSidebar from "./components/AdminSidebar"; // Ensure Sidebar is imported
 
 function App() {
   return (
@@ -20,8 +21,11 @@ function App() {
         <div className="flex flex-col h-screen">
           <Header />
           <div className="flex flex-1">
-            <Sidebar /> {/* Ensuring Sidebar is included */}
+            <UserSidebar /> {/* Ensuring Sidebar is included */}
             <div className="flex-1 overflow-auto p-4">
+            </div> {/* Closing the div to fix the lint error */}
+              <AdminSidebar />
+              <div className="flex-1 overflow-auto p-4">
               <Routes>
                 <Route path="/home" element={<HomePage />} />
                 <Route path="/login" element={<Login />} /> {/* Login page */}
